+++ examples/angle.hxx
@@ -1,3 +1,4 @@
+#pragma once
 
 //          Copyright David Browne 2020-2025.
 // Distributed under the Boost Software License, Version 1.0.