+++ examples/tolerance.hxx
@@ -1,3 +1,4 @@
+#pragma once
 
 //          Copyright David Browne 2020-2025.
 // Distributed under the Boost Software License, Version 1.0.
@@ -15,9 +16,14 @@ namespace dsga
 	//
 	// Basic tolerance comparisons:
 	// 
-	//	* within_tolerance(x, tolerance) - vectors treated as independent components for checks on x
+	//	* within_tolerance(x, tolerance) - vectors/matrices treated as independent components for checks on x
 	//	* within_distance(x, y, tolerance) - vectors treated as a whole for Euclidean distance check on x and y
-	//	* within_box(x, y, tolerance) - vectors treated as whole for orthogonal box check on x and y
+	//	* within_box(x, y, tolerance) - vectors/matrices treated as whole for orthogonal box check on x and y.
+	//
+	//	if you want component-wise checks, use "within_tolerance(x - y, tolerance)" for comparing 2 vectors.
+	//	For matrices, since we can't have a matrix of bools, we return a vector of bools, where each value
+	//	indicates whether a matrix column has all its elements within tolerance or not, and the tolerance
+	//	vector elements (assuming that tolerance is not a scalar) are tolerances for each entire column.
 	//
 
 	// General Tolerance Checking - checks if one or more values is close to 0 within the tolerances.
@@ -27,7 +33,7 @@ namespace dsga
 
 	template <floating_point_scalar T, floating_point_scalar U>
 	requires implicitly_convertible_to<U, T>
-	[[nodiscard]] constexpr bool within_tolerance(T x,
+	[[nodiscard]] constexpr auto within_tolerance(T x,
 												  U tolerance) noexcept
 	{
 		return abs(x) <= static_cast<T>(abs(tolerance));
@@ -75,6 +81,32 @@ namespace dsga
 		{
 			return within_tolerance(x, tolerance[0]);
 		}
+	}
+
+	// Tolerance checks for matrices - column-wise checks for matrices, i.e., each column is treated as a vector.
+	// returns a vector of boolean values, where each value indicates whether the corresponding column has all
+	// its elements within tolerance or not.
+
+	template <floating_point_scalar T, std::size_t C, std::size_t R, floating_point_scalar U>
+	requires implicitly_convertible_to<U, T>
+	[[nodiscard]] constexpr auto within_tolerance(const basic_matrix<T, C, R> &arg,
+												  U tolerance) noexcept
+	{
+		return [&arg, &tolerance] <std::size_t ...Is>(std::index_sequence<Is ...>) noexcept
+		{
+			return basic_vector<bool, C>(all(within_tolerance(arg[Is], tolerance)) ...);
+		}(std::make_index_sequence<C>{});
+	}
+
+	template <bool W, floating_point_scalar T, std::size_t C, std::size_t R, floating_point_scalar U, typename D>
+	requires implicitly_convertible_to<U, T>
+	[[nodiscard]] constexpr auto within_tolerance(const basic_matrix<T, C, R> &arg,
+												  const vector_base<W, U, C, D> &tolerance) noexcept
+	{
+		return [&arg, &tolerance] <std::size_t ...Is>(std::index_sequence<Is ...>) noexcept
+		{
+			return basic_vector<bool, C>(all(within_tolerance(arg[Is], tolerance[Is])) ...);
+		}(std::make_index_sequence<C>{});
 	}
 
 	// Euclidean distance check - imagine x in the center of a region (offset number line, circle, sphere, hypersphere)
@@ -148,4 +180,25 @@ namespace dsga
 		}
 	}
 
-}
+	// Tolerance checks for matrices - returns a boolean value indicating whether all the
+	// matrix elements are within the tolerance box .
+
+	template <floating_point_scalar T, std::size_t C, std::size_t R, floating_point_scalar U>
+	requires implicitly_convertible_to<U, T>
+	[[nodiscard]] constexpr bool within_box(const basic_matrix<T, C, R> &x,
+											const basic_matrix<T, C, R> &y,
+											U tolerance) noexcept
+	{
+		return all(within_tolerance(x - y, tolerance));
+	}
+
+	template <bool W, floating_point_scalar T, std::size_t C, std::size_t R, floating_point_scalar U, typename D>
+	requires implicitly_convertible_to<U, T>
+	[[nodiscard]] constexpr bool within_box(const basic_matrix<T, C, R> &x,
+											const basic_matrix<T, C, R> &y,
+											const vector_base<W, U, C, D> &tolerance) noexcept
+	{
+		return all(within_tolerance(x - y, tolerance));
+	}
+
+}	// namespace dsga